+++ Entidades/Pagos.cs
@@ -7,7 +7,7 @@ namespace Entidades
         public int IdPagos { get; set; }
         public int IdCuota { get; set; }
         public DateTime FechaPago { get; set; }
-        public int MontoPagado { get; set; }
+        public decimal MontoPagado { get; set; }
         public int? MoraAplicada { get; set; }
     }
 }
